fix(webpush): Move enums to dedicated files to avoid future autoloader issues
