+++ lib/WebPush/TokenValidation.php
@@ -7,7 +7,7 @@ declare(strict_types=1);
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
-namespace OCA\Notifications;
+namespace OCA\Notifications\WebPush;
 
 enum TokenValidation: int {
 	case VALID = 0;
