+++ tests/Unit/Controller/WebPushControllerTest.php
@@ -12,9 +12,9 @@ use OC\Authentication\Exceptions\InvalidTokenException;
 use OC\Authentication\Token\IProvider;
 use OC\Authentication\Token\IToken;
 use OC\Security\IdentityProof\Manager;
-use OCA\Notifications\Controller\ActivationSubStatus;
-use OCA\Notifications\Controller\NewSubStatus;
 use OCA\Notifications\Controller\WebPushController;
+use OCA\Notifications\WebPush\ActivationSubStatus;
+use OCA\Notifications\WebPush\NewSubStatus;
 use OCA\Notifications\WebPushClient;
 use OCP\AppFramework\Http;
 use OCP\AppFramework\Http\DataResponse;
